+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ### Added
 
-- feat: replace templates with gomponents
+- feat: add go template functions to render a form, a field, etc.
 - feat: add constraints (isodd, iseven)
+
+### Changed
+
+- feat: replace templates with gomponents
 
 ## v1.1.0